concierge_service_frontend/pyproject.toml
[project]
name = "concierge_service_frontend"
version = "0.1.0"
description = ""
authors = [
    {name = "Shubham",email = "shubhammahawar.26@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0 !=3.9.7"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
